+++ r-postgresql.tf
@@ -44,9 +44,9 @@ resource "azurerm_postgresql_database" "postgresql_db" {
 }
 
 resource "azurerm_postgresql_configuration" "postgresql_config" {
-  count               = length(var.postgresql_configurations)
-  name                = var.postgresql_configurations[count.index].name
+  for_each            = var.postgresql_configurations
+  name                = each.key
   resource_group_name = var.resource_group_name
   server_name         = azurerm_postgresql_server.postgresql_server.name
-  value               = var.postgresql_configurations[count.index].value
+  value               = each.value
 }